Map can be used in place of POJO added method to get placeholder set.
