+++ .gitignore
@@ -1,3 +1,9 @@
+# Compiled class file
+*.class
+
+# Log file
+*.log
+
 .idea
 *.iml
 settings.xml
@@ -6,3 +12,16 @@ settings.xml
 /.DS_Store
 /.project
 /.gradle/
+
+# Package Files #
+*.jar
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+/bin/
